pyproject.toml
[project]
name = "TACA"
version = "1.0.0"
description = "TACA"
license = "MIT"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = "==3.11.*"
dependencies = [
    "anndata==0.10.6",
    "pandas==2.2.1",
    "scanpy==1.9.8",
    "matplotlib==3.8.3",
    "scvi-tools==1.1.2",
    "numpy==1.26.4",
    "umap-learn==0.5.5",
    "scikit-learn==1.4.1.post1",
    "optuna==4.4.0",
    "seaborn==0.13.2",
    "annoy==1.17.2",
    "scrublet==0.2.3",
    "scipy==1.12.0"
]


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = []  # no packages; we just want to install dependencies
# (equivalently)
# [tool.setuptools.packages.find]
# include = []
# exclude = ["*"]
